+++ src/backend/commands/dbmigrate.command.ts
@@ -1,9 +1,19 @@
-import {AbstractCommand} from '@dps/mycms-server-commons/dist/backend-commons/commands/abstract.command';
+import {CommonAdminCommand, SimpleConfigFilePathValidationRule, SimpleFilePathValidationRule} from './common-admin.command';
+import {NameValidationRule, ValidationRule} from '@dps/mycms-commons/dist/search-commons/model/forms/generic-validator.util';
+
 const DBMigrate = require('db-migrate');
 
 // TODO move to commons
-export class DbMigrateCommand implements AbstractCommand {
-    public process(argv): Promise<any> {
+export class DbMigrateCommand extends CommonAdminCommand {
+    protected createValidationRules(): {[key: string]: ValidationRule} {
+        return {
+            migrationDbConfigFile: new SimpleConfigFilePathValidationRule(true),
+            migrationsDir: new SimpleFilePathValidationRule(true),
+            migrationEnv: new NameValidationRule(true)
+        };
+    }
+
+    protected processCommandArgs(argv: {}): Promise<any> {
         const migrationDbConfigFile = argv['migrationDbConfigFile'];
         if (migrationDbConfigFile === undefined) {
             return Promise.reject('ERROR - parameters required migrationDbConfigFile: "--migrationDbConfigFile"');