+++ src/backend/commands/sitemap-generator.command.ts
@@ -4,19 +4,27 @@ import {PDocSearchForm} from '@dps/mycms-commons/dist/pdoc-commons/model/forms/p
 import {PDocRecord} from '@dps/mycms-commons/dist/pdoc-commons/model/records/pdoc-record';
 import {TourDocDataServiceModule} from '../modules/tdoc-dataservice.module';
 import {PDocDataServiceModule} from '../modules/pdoc-dataservice.module';
-import {AbstractCommand} from '@dps/mycms-server-commons/dist/backend-commons/commands/abstract.command';
 import {CommonDocSearchForm} from '@dps/mycms-commons/dist/search-commons/model/forms/cdoc-searchform';
 import {CommonDocRecord} from '@dps/mycms-commons/dist/search-commons/model/records/cdoc-entity-record';
 import {CommonDocSearchResult} from '@dps/mycms-commons/dist/search-commons/model/container/cdoc-searchresult';
 import {CommonDocDataService} from '@dps/mycms-commons/dist/search-commons/services/cdoc-data.service';
 import {StringUtils} from '@dps/mycms-commons/dist/commons/utils/string.utils';
+import {CommonAdminCommand, SimpleConfigFilePathValidationRule} from './common-admin.command';
+import {ValidationRule} from '@dps/mycms-commons/dist/search-commons/model/forms/generic-validator.util';
 
-export class SiteMapGeneratorCommand implements AbstractCommand {
-    public process(argv): Promise<any> {
-        const filePathConfigJson = argv['c'] || argv['backend'];
-        const filePathSitemapConfigJson = argv['s'] || argv['sitemap'] || argv['_'][0];
+export class SiteMapGeneratorCommand extends CommonAdminCommand {
+    protected createValidationRules(): {[key: string]: ValidationRule} {
+        return {
+            backend: new SimpleConfigFilePathValidationRule(true),
+            sitemap: new SimpleConfigFilePathValidationRule(true)
+        };
+    }
+
+    protected processCommandArgs(argv: {}): Promise<any> {
+        const filePathConfigJson = argv['backend'];
+        const filePathSitemapConfigJson = argv['sitemap'];
         if (filePathConfigJson === undefined || filePathSitemapConfigJson === undefined) {
-            return Promise.reject('ERROR - parameters required backendConfig: "-c | --backend" sitemapConfig: "-s | --sitemap"');
+            return Promise.reject('ERROR - parameters required backendConfig: "--backend" sitemapConfig: "--sitemap"');
         }
 
         const generatorConfig = {