+++ src/backend/commands/tdoc-exporter.command.ts
@@ -1,4 +1,3 @@
-import {AbstractCommand} from '@dps/mycms-server-commons/dist/backend-commons/commands/abstract.command';
 import * as fs from 'fs';
 import {TourDocDataServiceModule} from '../modules/tdoc-dataservice.module';
 import {TourDocAdapterResponseMapper} from '../shared/tdoc-commons/services/tdoc-adapter-response.mapper';
@@ -10,15 +9,24 @@ import {GenericAdapterResponseMapper} from '@dps/mycms-commons/dist/search-commo
 import {CommonDocTransportModule} from '@dps/mycms-server-commons/dist/backend-commons/modules/cdoc-transport.module';
 import {utils} from 'js-data';
 import {TourDocFileUtils} from '../shared/tdoc-commons/services/tdoc-file.utils';
+import {CommonAdminCommand, SimpleConfigFilePathValidationRule, SimpleFilePathValidationRule} from './common-admin.command';
+import {ValidationRule} from '@dps/mycms-commons/dist/search-commons/model/forms/generic-validator.util';
 
-export class TourDocExporterCommand implements AbstractCommand {
-    public process(argv): Promise<any> {
+export class TourDocExporterCommand extends CommonAdminCommand {
+    protected createValidationRules(): {[key: string]: ValidationRule} {
+        return {
+            backend: new SimpleConfigFilePathValidationRule(true),
+            file: new SimpleFilePathValidationRule(true)
+        };
+    }
+
+    protected processCommandArgs(argv: {}): Promise<any> {
         const perRun = 999;
         const typeOrder = ['location', 'news', 'trip', 'route', 'track', 'image', 'video'];
 
-        const filePathConfigJson = argv['c'] || argv['backend'];
+        const filePathConfigJson = argv['backend'];
         if (filePathConfigJson === undefined) {
-            return Promise.reject('ERROR - parameters required backendConfig: "-c | --backend"');
+            return Promise.reject('ERROR - parameters required backendConfig: "--backend"');
         }
 
         const serverConfig = {
@@ -26,7 +34,7 @@ export class TourDocExporterCommand implements AbstractCommand {
             readOnly: false
         };
 
-        const dataFileName = TourDocFileUtils.normalizeCygwinPath(argv['f'] || argv['file']);
+        const dataFileName = TourDocFileUtils.normalizeCygwinPath(argv['file']);
         if (dataFileName === undefined) {
             return Promise.reject('option --file expected');
         }